Refactor LLM service route
