+++ usellm.org/app/api/llm/route.ts
@@ -1,5 +1,47 @@
-import { llmService } from "@/lib/llm-service";
-import { getChatResponseHeaders } from "@/lib/utils";
+import { Ratelimit } from "@upstash/ratelimit";
+import { Redis } from "@upstash/redis";
+import { createLLMService } from "usellm";
+
+export const llmService = createLLMService({
+  openaiApiKey: process.env.OPENAI_API_KEY,
+  elvenLabsApiKey: process.env.ELVEN_LABS_API_KEY,
+  actions: [
+    "chat",
+    "transcribe",
+    "embed",
+    "speak",
+    "generateImage",
+    "editImage",
+    "imageVariation",
+  ],
+  isAllowed: async () => {
+    // check if rate limiting has been set up using Upstash Redis REST API
+    if (
+      process.env.UPSTASH_REDIS_REST_URL &&
+      process.env.UPSTASH_REDIS_REST_TOKEN
+    ) {
+      const { success } = await ratelimit.limit("api");
+      return success;
+    }
+    return true;
+  },
+});
+
+llmService.registerTemplate({
+  id: "jobot",
+  systemPrompt:
+    "Your name is Jobot! You have been developed by Jovian to help the world.",
+  userPrompt: "Tell me about {{topic}}",
+});
+
+llmService.registerTemplate({
+  id: "tutorial-generator",
+  systemPrompt:
+    "You job is to create a short tutorial on a given topic. Use simple words, avoid jargon. Start with an introduction, then provide a few points of explanation, and end with a conclusion",
+  userPrompt: "Topic: {{topic}}",
+  max_tokens: 200,
+  temperature: 0.7,
+});
 
 export const runtime = "edge";
 
@@ -40,3 +82,19 @@ export async function POST(request: Request) {
     });
   }
 }
+
+function getChatResponseHeaders() {
+  return {
+    "Access-Control-Allow-Credentials": "true",
+    "Access-Control-Allow-Origin": "*",
+    "Access-Control-Allow-Methods": "GET,OPTIONS,PATCH,DELETE,POST,PUT",
+    "Content-Type": "application/json",
+    "Access-Control-Allow-Headers":
+      "X-CSRF-Token, X-Requested-With, Accept, Accept-Version, Content-Length, Content-MD5, Content-Type, Date, X-Api-Version, X-Referer, Authorization, API_URL",
+  };
+}
+
+const ratelimit = new Ratelimit({
+  redis: Redis.fromEnv(),
+  limiter: Ratelimit.slidingWindow(10, "10 s"),
+});
